refactor: finish all useful refactoring
