+++ src/hooks/helpers/reduxHooks.tsx
@@ -4,7 +4,7 @@ import {
   useSelector,
 } from 'react-redux';
 
-import type { AppDispatch, RootState } from '../app/AppStore';
+import type { AppDispatch, RootState } from '../../app/AppStore';
 
 export const useAppDispatch = useDispatch<AppDispatch>;
 export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
