+++ src/hooks/helpers/useProductsService.tsx
@@ -2,16 +2,17 @@
 import { useEffect } from 'react';
 import { useAppDispatch, useAppSelector } from './reduxHooks';
 import useIsStaffPage from './useIsStaffPage';
-import { resetState, setQuery } from '../app/reducers/ProductsSlice';
-import { useGetProductsQuery } from '../app/api/productsApi';
+import { resetState, setQuery } from '../../app/reducers/ProductsSlice';
+import { useGetProductsQuery } from '../../app/api/productsApi';
 import { useLocation } from 'react-router-dom';
 
 export const useProductsService = () => {
   const dispatch = useAppDispatch();
-
-  const isStaffPage = useIsStaffPage();
+  const location = useLocation();
   const { skip, selectedCategoryQuery, query, limit, searchQuery } =
     useAppSelector(state => state.products);
+
+  const isStaffPage = useIsStaffPage();
 
   const {
     data: productsData,
@@ -21,15 +22,14 @@ export const useProductsService = () => {
 
   const getInitialQuery = () => {
     if (isStaffPage) {
-      if (searchQuery !== '' && searchQuery !== undefined) {
+      const isSearchQuery = searchQuery !== '' && searchQuery !== undefined;
+      if (isSearchQuery) {
         return { search: searchQuery };
       }
       return { limit, skip };
     }
     return { category: selectedCategoryQuery, limit, skip };
   };
-
-  const location = useLocation();
 
   useEffect(() => {
     const resultQuery = getInitialQuery();
@@ -41,10 +41,6 @@ export const useProductsService = () => {
     const resultQuery = getInitialQuery();
     dispatch(setQuery(resultQuery));
   }, [selectedCategoryQuery, skip, dispatch, searchQuery]);
-
-  useEffect(() => {
-    console.log(query);
-  }, [query]);
 
   return {
     productsData,
