+++ src/hooks/useSearch.tsx
@@ -1,5 +1,5 @@
-import { useEffect, useRef } from 'react';
-import { useAppDispatch, useAppSelector } from './reduxHooks';
+import { useRef } from 'react';
+import { useAppDispatch, useAppSelector } from './helpers/reduxHooks';
 import { setSearchQuery } from '../app/reducers/ProductsSlice';
 
 import { debounce } from 'lodash';
@@ -12,7 +12,6 @@ export const useSearch = () => {
   const debouncedSetSearchRef = useRef(
     debounce(value => {
       dispatch(setSearchQuery(value));
-      console.log('debounce');
     }, 700)
   );
 
@@ -28,13 +27,8 @@ export const useSearch = () => {
   const onKeyPress = (e: React.KeyboardEvent<HTMLInputElement>) => {
     if (e.key === 'Enter') {
       debouncedSetSearchRef.current.flush();
-      console.log('Enter');
     }
   };
-
-  useEffect(() => {
-    console.log('query', searchQuery);
-  }, [searchQuery]);
 
   return {
     handleSearch,
